# MM-63756: Added index to sidebarchannels table