@agnivade 2025-04-15
> One non-blocking thought: probably not worth it given team ids generally have low cardinality, but I think we don't have an index on sidebarcategories.teamid either.

Do you have a query which can be improved from this index?

> Also, since changes are contained, thoughts on backporting this to the latest ESR?

Typically we try to avoid schema changes in dot releases. I'd say if this becomes a widespread issue, then happy to backport.