@agnivade 2025-04-15
> I am mostly wondering whether we think the queries in question could benefit from that.

Ah no, I checked them. The current index is good enough for that.

> Well, we have customers potentially struggling with this today. Something to keep in mind.

Yep. Only one customer (so far) will actually benefit from this. Will keep an eye. Otherwise, happy to backport.