@mattermost-build 2025-05-05
Error trying doing the automated Cherry picking. Please do this manually

```
+++ Updating remotes...
Fetching upstream
hostfile_replace_entries: mkstemp: Read-only file system
update_known_hosts: hostfile_replace_entries failed for /app/.ssh/known_hosts: Read-only file system
From github.com:mattermost/mattermost
   0dac5f532a..223e329e05  MM-62751              -> upstream/MM-62751
   5a03f6270a..9865555291  MM-63837-net_version  -> upstream/MM-63837-net_version
 * [new branch]            MM-63925_sidebar-categories-transaction -> upstream/MM-63925_sidebar-categories-transaction
 * [new branch]            MM-63950-Replace-Denim-Button-Variables -> upstream/MM-63950-Replace-Denim-Button-Variables
 * [new branch]            MM-63950-Use-Proper-Button-Classes-No-Denim -> upstream/MM-63950-Use-Proper-Button-Classes-No-Denim
 * [new branch]            accessibility__browse_channels -> upstream/accessibility__browse_channels
 * [new branch]            accessibility__drafts -> upstream/accessibility__drafts
 * [new branch]            backport-playwright-shared-library -> upstream/backport-playwright-shared-library
 * [new branch]            consistent-button-classes -> upstream/consistent-button-classes
   92df337cbd..f4735d6513  dev-abac              -> upstream/dev-abac
   1e69a9cda2..9622c17764  e2e-test-fix-scheduled-posts -> upstream/e2e-test-fix-scheduled-posts
 * [new branch]            mm-62158-group-store-no-select-star-part2 -> upstream/mm-62158-group-store-no-select-star-part2
 * [new branch]            mm-62158-group-store-no-select-star-part3 -> upstream/mm-62158-group-store-no-select-star-part3
   29b1c7571a..b7040a99e6  mm-63406-timezone-auto-changes-focus-timer -> upstream/mm-63406-timezone-auto-changes-focus-timer
   5bd1836572..ca124effd8  release-10.8          -> upstream/release-10.8
Fetching upstream
hostfile_replace_entries: mkstemp: Read-only file system
update_known_hosts: hostfile_replace_entries failed for /app/.ssh/known_hosts: Read-only file system
+++ Updating remotes done...
+++ Creating local branch automated-cherry-pick-of-mattermost-#30724-upstream-release-9.11-1746414615
Switched to a new branch 'automated-cherry-pick-of-mattermost-#30724-upstream-release-9.11-1746414615'
Branch 'automated-cherry-pick-of-mattermost-#30724-upstream-release-9.11-1746414615' set up to track remote branch 'release-9.11' from 'upstream'.

+++ About to attempt cherry pick of PR #30724 with merge commit 131cf039bbd5d257e1bcb8ceeca6bc48eef8b0a4.

Auto-merging server/channels/db/migrations/migrations.list
CONFLICT (content): Merge conflict in server/channels/db/migrations/migrations.list
error: could not apply 131cf039bb... MM-63756: Added index to sidebarchannels table (#30724)
hint: After resolving the conflicts, mark them with
hint: "git add/rm <pathspec>", then run
hint: "git cherry-pick --continue".
hint: You can instead skip this commit with "git cherry-pick --skip".
hint: To abort and get back to the state before "git cherry-pick",
hint: run "git cherry-pick --abort".

+++ Conflicts detected:

UU server/channels/db/migrations/migrations.list
Aborting.

+++ Aborting in-progress git cherry-pick.

+++ Returning you to the master branch and cleaning up.

```