@agnivade 2025-05-05
/cherry-pick release-10.5