@mattermost-build 2025-05-05
Error trying doing the automated Cherry picking. Please do this manually

```
+++ Updating remotes...
Fetching upstream
hostfile_replace_entries: mkstemp: Read-only file system
update_known_hosts: hostfile_replace_entries failed for /app/.ssh/known_hosts: Read-only file system
Fetching upstream
hostfile_replace_entries: mkstemp: Read-only file system
update_known_hosts: hostfile_replace_entries failed for /app/.ssh/known_hosts: Read-only file system
+++ Updating remotes done...
+++ Creating local branch automated-cherry-pick-of-mattermost-#30724-upstream-release-10.5-1746414644
Switched to a new branch 'automated-cherry-pick-of-mattermost-#30724-upstream-release-10.5-1746414644'
Branch 'automated-cherry-pick-of-mattermost-#30724-upstream-release-10.5-1746414644' set up to track remote branch 'release-10.5' from 'upstream'.

+++ About to attempt cherry pick of PR #30724 with merge commit 131cf039bbd5d257e1bcb8ceeca6bc48eef8b0a4.

Auto-merging server/channels/db/migrations/migrations.list
CONFLICT (content): Merge conflict in server/channels/db/migrations/migrations.list
error: could not apply 131cf039bb... MM-63756: Added index to sidebarchannels table (#30724)
hint: After resolving the conflicts, mark them with
hint: "git add/rm <pathspec>", then run
hint: "git cherry-pick --continue".
hint: You can instead skip this commit with "git cherry-pick --skip".
hint: To abort and get back to the state before "git cherry-pick",
hint: run "git cherry-pick --abort".

+++ Conflicts detected:

UU server/channels/db/migrations/migrations.list
Aborting.

+++ Aborting in-progress git cherry-pick.

+++ Returning you to the master branch and cleaning up.

```